# llama : make quantize example up to 2.7x faster