Tested on both ramfs and NVMe-backed btrfs. My CPU is a 6-core, 12-thread Ryzen 5 3600.

(These numbers are out of date due to some changes being removed, see the discussion. The final speedup for this PR is about 2.7x at best.)

| model | format | cache | master       | PR          | speedup |
| ----: | - | ----- | -----------: | ----------: | ------- |
| 7B    |Q4_0| ramfs | 19054.03 ms  | 5176.92 ms  | 3.68    |
| 33B   |Q4_0| cold  | 127157.88 ms | 44832.70 ms | 2.84    |
| 33B   |Q4_0| warm  | 126347.77 ms | 44401.31 ms | 2.85    |