@Cebtenzzre 2023-09-10
Oops, I forgot that Windows doesn't have pread.