@bobqianic 2023-09-11
> What is this black magic !?!?

https://github.com/ggerganov/llama.cpp/blob/f31b6f4e2d6def3c0bd7c75f75c0c1e8698e0589/llama.cpp#L4950

I think the primary improvement is the relocation of `f32_conv_buf`. Previously, `f32_conv_buf` was inside the loop, but now it's outside the loop. This way, there's no need to continuously allocate and deallocate memory.