@ikawrakow 2023-09-11
This PR makes expensive quantizations (k_quants) slower. E.g., for `Q4_K_S` I get 47.4 seconds with this PR vs 43.4 seconds on Master. When it comes to already fast quantizations (`Q4_0`, `Q4_1`), I'm not really bothered by the fact that I have to weight for 5 seconds (instead of 2.5 seconds with this PR) for the quantization to finish.