@ikawrakow 2023-09-11
> edit: 5 seconds? Quantizing 33B to Q4_0 on master takes two minutes on my hardware. Am I doing something wrong?

The numbers I gave in the comment were for 7B. For 33B I have 10 seconds (Master) vs 4.2 seconds (PR) for `Q4_0`, and 83 seconds (Master) vs 97 seconds (PR) for `Q4_K_S`. This is on M2 Max.