@Cebtenzzre 2023-09-11
> For 30B and `Q4_0` I have 35.7 seconds on master and ~60 seconds with this PR.

I suppose I'll have to do some tuning on different platforms. It seems like what applies to Linux on amd64 may not apply to macOS on aarch64, or Windows for that matter. One thing you could try is clearing your disk cache before each run - that's where the extra parallelization should help.