@ikawrakow 2023-09-12
With the latest version (46d1b6e1cdcf41d959b2df4e480f1c9fb29d5369) it behaves better. I get 28 seconds for 30B and `Q4_0` without it being choppy as before. But I get the exact same 28 seconds with the master version by moving `f32_conv_buf` out of the loop over tensors as @bobqianic suggested above.

For 30B and `Q4_K_S` I get 219.4 seconds with this PR vs 215.7 seconds on master with  `f32_conv_buf`  moved out of the loop.