@ikawrakow 2023-09-12
Oh, and if I disable quant histogram collection, then for `Q4_K_S` I'm down to 2.7 seconds for 7B and 16 seconds for 30B. This is on master with `f32_conv_buf` moved out of the loop.