@ikawrakow 2023-09-12
@Cebtenzzre 

I pushed a very simple modification to the quantization to https://github.com/ggerganov/llama.cpp/tree/ik/quantize_faster. It just changes two things:
* Moved `f32_conv_buf` out of the loop over tensors
* Added the ability to enable/disable quant histogram collection.

With quant histogram collection enabled, your PR is about the same as https://github.com/ggerganov/llama.cpp/tree/ik/quantize_faster on my M2 Max, but slower on a Ryzen 7950X (e.g., 43.7 seconds vs 31.9 seconds for 30B and `Q4_0`). With quant histogram collection disabled, https://github.com/ggerganov/llama.cpp/tree/ik/quantize_faster is faster than your PR by a large margin. I'm curious to know how https://github.com/ggerganov/llama.cpp/tree/ik/quantize_faster performs on your system. The command line is
```
./quantize $model $output_file $quant_type $num_threads $enable_histo
```
where the last argument `$enable_histo` is 0 or 1, with 0 disabling and 1 enabling histogram collection.