@Cebtenzzre 2023-09-12
7B source on ramfs, dest is /dev/null
| model | quant  | master   | Old PR       | PR       | PR+mmap      | ikawrakow |
|------:|--------|---------:|-------------:|---------:|-------------:|----------:|
| 7B    | Q4_0   | 20.45 s  | 5.33 s       | 6.63 s   | **4.74 s**   | 7.89 s    |
| 7B    | Q4_K_S | 83.28 s  | **67.00 s**  | 69.02 s  | **67.82 s**  | 69.73 s   |

33B source on NVMe, dest on NVMe, deleted after each run
cache is cleared before each run with `sync; echo 3 > /proc/sys/vm/drop_caches`
| model | quant  | master   | Old PR       | PR       | PR+mmap      | ikawrakow |
|------:|--------|---------:|-------------:|---------:|-------------:|----------:|
| 33B   | Q4_0   | 153.29 s | **63.11 s**  | 93.98 s  | 82.22 s      | 63.38 s   |
| 33B   | Q4_K_S | 445.30 s | **367.33 s** | 385.64 s | **368.34 s** | 387.69 s  |

"Old PR" is commit 96c8042884196f6e23fcfbf5470aa6df05fd4db3.
"PR" is commit 46d1b6e1cdcf41d959b2df4e480f1c9fb29d5369.
"PR+mmap" is 46d1b6e1cdcf41d959b2df4e480f1c9fb29d5369 with use_mmap set to true.
"ikawrakow" is commit da030ed0635e38717358d45f8a322b22ac0814f5 with histograms enabled for comparison's sake.
"Old PR" and "PR+mmap" are tied for k-quants because they are essentially doing the same thing.

All options are faster than master for me, so for the time being I'd be happy with any of them. I wouldn't want to hurt performance for anyone.