@ikawrakow 2023-09-13
Quantization time in seconds on M2 Max and Ryzen 7950X for `Q4_0`:

| model, platform | This PR | This PR + mmap | ikawrakow | ikawrakow - no histo |
|---|---|---|---|---|
| 33 B, M2 Max | 28.0 | 66.3 | 28.0 | 16.0 |
| 33B, 7950X | 43.7 | 27.9 | 31.9 | 26.9 |

So, mmap does help this PR on Linux, making it ~14% faster than the one-line change that is labeled "ikawrakow".

But using mmap on my 64 GB RAM M2 Max is a disaster (I guess, @ggerganov does not have the issue with 192 GB of RAM on his M2 Ultra). With mmap disabled, the one-line change clearly wins.

If the consensus is still somehow that the +241-139 lines change in this PR is better compared to the +1-1 line change, then let's at least add the ability to disable/enable quant histogram collection as in the https://github.com/ggerganov/llama.cpp/tree/ik/quantize_faster branch. I never look at these histograms (but whoever wants to look at them will still be able to) and, as can be seen in the last column of the above table, disabling quant histogram collection brings a significant speedup (well, at least when it comes to quantizing `Q4_0`, `Q4_1`, `Q5_0`, `Q5_1`, `Q8_0`).