@Cebtenzzre 2023-09-13
> Quantization time in seconds on M2 Max and Ryzen 7950X for `Q4_0`:

I wish I could understand what makes this PR slower on your 7950X. Would you mind stepping through the commits one-by-one, disabling use_mmap on the commits where it is true, and telling me which one brings the regression relative to master?

I'm all for avoiding unnecessary code changes, but I don't see why we should ignore any obvious opportunities for optimization, either.